Break vecFunctions into separate files
