+++ src/vecFunctions/vecAdd.ts
@@ -1,0 +1,14 @@
+import { IVec } from "../types";
+import { vecAlloc } from "./vecAlloc";
+import { vecReset } from "./vecReset";
+
+/**
+ * Computes the result of adding the two given vectors.
+ *
+ * @param a
+ * @param b
+ * @param out
+ */
+export function vecAdd(a: IVec, b: IVec, out = vecAlloc()) {
+  return vecReset(a.x + b.x, a.y + b.y, out);
+}
