+++ src/vecFunctions/vecAlloc.ts
@@ -1,0 +1,15 @@
+import { _vecAlloc } from "../internal/dataClasses";
+import { IVec } from "../types";
+
+class Vec implements IVec {
+  constructor(public x = NaN, public y = NaN) {}
+}
+
+/**
+ * Creates a new Vec object in memory, with all values initialized to NaN.
+ * This is useful to hold the result of vectormath function calls in performance
+ * critical workflows.
+ */
+export function vecAlloc(): IVec {
+  return new Vec();
+}
